Protection ToGithub & FromGithub fixed
For #60 - fixed protection.ToGithub and protection.FromGithub.

Get protection command returned not empty Review.Dismiss struct if dismissal users and teams was not specified. So it was marshalled as "dismiss: {}", and couldn't be omited.

Apply protection command did not pass empty DismissalRestrictionsRequest if users and teams was not specified. Therefore it was not possible to disable dismissal restictions. It could only empy lists of users and teams.

In both cases added check for emptiness of users and teams lists.

PR: #73